update according to comments
