+++ pkg/istio/rbac/common/servicerole_test.go
@@ -9,7 +9,6 @@ import (
 
 	"istio.io/api/rbac/v1alpha1"
 
-	"github.com/yahoo/k8s-athenz-istio-auth/pkg/istio/common"
 	"github.com/stretchr/testify/assert"
 	"github.com/yahoo/athenz/clients/go/zms"
 	"github.com/yahoo/k8s-athenz-istio-auth/pkg/log"
@@ -17,178 +16,6 @@ import (
 
 func init() {
 	log.InitLogger("", "debug")
-}
-
-func TestParseAssertionEffect(t *testing.T) {
-
-	allow := zms.ALLOW
-	deny := zms.DENY
-	cases := []struct {
-		test           string
-		assertion      *zms.Assertion
-		expectedEffect string
-		expectedErr    error
-	}{
-		{
-			test:           "empty assertion",
-			assertion:      nil,
-			expectedEffect: "",
-			expectedErr:    fmt.Errorf("assertion is nil"),
-		},
-		{
-			test:           "empty assertion effect",
-			assertion:      &zms.Assertion{},
-			expectedEffect: "",
-			expectedErr:    fmt.Errorf("assertion effect is nil"),
-		},
-		{
-			test: "valid effect",
-			assertion: &zms.Assertion{
-				Effect: &allow,
-			},
-			expectedEffect: "ALLOW",
-			expectedErr:    nil,
-		},
-		{
-			test: "invalid(unsupported) effect",
-			assertion: &zms.Assertion{
-				Effect: &deny,
-			},
-			expectedEffect: "",
-			expectedErr:    fmt.Errorf("effect: DENY is not a supported assertion effect"),
-		},
-	}
-
-	for _, c := range cases {
-		gotAssertion, gotErr := common.ParseAssertionEffect(c.assertion)
-		assert.Equal(t, c.expectedEffect, gotAssertion, c.test)
-		assert.Equal(t, c.expectedErr, gotErr, c.test)
-	}
-}
-
-func TestParseAssertionAction(t *testing.T) {
-
-	cases := []struct {
-		test           string
-		assertion      *zms.Assertion
-		expectedAction string
-		expectedErr    error
-	}{
-		{
-			test:           "empty assertion",
-			assertion:      nil,
-			expectedAction: "",
-			expectedErr:    fmt.Errorf("assertion is nil"),
-		},
-		{
-			test: "valid action",
-			assertion: &zms.Assertion{
-				Action: "get",
-			},
-			expectedAction: "GET",
-			expectedErr:    nil,
-		},
-		{
-			test: "valid action POST",
-			assertion: &zms.Assertion{
-				Action: "POST",
-			},
-			expectedAction: "POST",
-			expectedErr:    nil,
-		},
-		{
-			test: "valid action wildcard *",
-			assertion: &zms.Assertion{
-				Action: "*",
-			},
-			expectedAction: "*",
-			expectedErr:    nil,
-		},
-		{
-			test: "invalid action",
-			assertion: &zms.Assertion{
-				Action: "launch",
-			},
-			expectedAction: "",
-			expectedErr:    fmt.Errorf("method: launch is not a supported HTTP method"),
-		},
-	}
-
-	for _, c := range cases {
-		gotAssertion, gotErr := common.ParseAssertionAction(c.assertion)
-		assert.Equal(t, c.expectedAction, gotAssertion, c.test)
-		assert.Equal(t, c.expectedErr, gotErr, c.test)
-	}
-}
-
-func TestParseAssertionResource(t *testing.T) {
-
-	cases := []struct {
-		test         string
-		domainName   zms.DomainName
-		assertion    *zms.Assertion
-		expectedSvc  string
-		expectedPath string
-		expectedErr  error
-	}{
-		{
-			test:         "empty assertion",
-			domainName:   "athenz.domain",
-			assertion:    nil,
-			expectedSvc:  "",
-			expectedPath: "",
-			expectedErr:  fmt.Errorf("assertion is nil"),
-		},
-		{
-			test:       "valid resource spec",
-			domainName: "athenz.domain",
-			assertion: &zms.Assertion{
-				Resource: "athenz.domain:svc.my-backend-service:/protected/endpoint",
-			},
-			expectedSvc:  "my-backend-service",
-			expectedPath: "/protected/endpoint",
-			expectedErr:  nil,
-		},
-		{
-			test:       "resource specifying valid service without endpoint",
-			domainName: "athenz.domain",
-			assertion: &zms.Assertion{
-				Resource: "athenz.domain:svc.my-backend-service",
-			},
-			expectedSvc:  "my-backend-service",
-			expectedPath: "",
-			expectedErr:  nil,
-		},
-		{
-			test:       "resource specifying service of another domain",
-			domainName: "athenz.domain",
-			assertion: &zms.Assertion{
-				Resource: "some.other.athenz.domain:svc.my-backend-service:/protected/endpoint",
-			},
-			expectedSvc:  "",
-			expectedPath: "",
-			expectedErr: fmt.Errorf("resource: some.other.athenz.domain:svc.my-backend-service:/protected/endpoint" +
-				" does not belong to the Athenz domain: athenz.domain"),
-		},
-		{
-			test:       "resource not specifying a service in required format",
-			domainName: "athenz.domain",
-			assertion: &zms.Assertion{
-				Resource: "athenz.domain:service.my-backend-service:/protected/endpoint",
-			},
-			expectedSvc:  "",
-			expectedPath: "",
-			expectedErr: fmt.Errorf("resource: athenz.domain:service.my-backend-service:/protected/endpoint does " +
-				"not specify the service using svc.<service-name> format"),
-		},
-	}
-
-	for _, c := range cases {
-		gotSvc, gotPath, gotErr := common.ParseAssertionResource(c.domainName, c.assertion)
-		assert.Equal(t, c.expectedSvc, gotSvc, c.test)
-		assert.Equal(t, c.expectedPath, gotPath, c.test)
-		assert.Equal(t, c.expectedErr, gotErr, c.test)
-	}
 }
 
 func TestConvertAthenzRoleNameToK8sName(t *testing.T) {
