--- pkg/istio/rbac/v2/provider_test.go
@@ -1,176 +1,0 @@
-package v2
-
-import (
-	"github.com/ardielle/ardielle-go/rdl"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/yahoo/athenz/clients/go/zms"
-	m "github.com/yahoo/k8s-athenz-istio-auth/pkg/athenz"
-	fakev1 "github.com/yahoo/k8s-athenz-syncer/pkg/client/clientset/versioned/fake"
-	adInformer "github.com/yahoo/k8s-athenz-syncer/pkg/client/informers/externalversions/athenz/v1"
-	"istio.io/api/security/v1beta1"
-	workloadv1beta1 "istio.io/api/type/v1beta1"
-	"istio.io/istio/pilot/pkg/model"
-	"istio.io/istio/pkg/config/schema/collections"
-	v1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/tools/cache"
-	"testing"
-	"time"
-)
-
-const (
-	DomainName = "test.namespace"
-	username   = "user.name"
-	username1  = "user.*"
-)
-
-var (
-	onboardedService = &v1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "onboarded-service",
-			Namespace: "test-namespace",
-			Annotations: map[string]string{
-				"authz.istio.io/enabled": "true",
-			},
-			Labels: map[string]string{
-				"app": "productpage",
-			},
-		},
-	}
-)
-
-func TestConvertAthenzModelIntoIstioAuthzPolicy(t *testing.T) {
-	athenzclientset := fakev1.NewSimpleClientset()
-	fakeAthenzInformer := adInformer.NewAthenzDomainInformer(athenzclientset, 0, cache.Indexers{})
-	signedDomain := getFakeDomain()
-	labels := onboardedService.GetLabels()
-	domainRBAC := m.ConvertAthenzPoliciesIntoRbacModel(signedDomain.Domain, &fakeAthenzInformer)
-
-	p := NewProvider(true)
-	
-	convertedCR := p.ConvertAthenzModelIntoIstioAuthzPolicy(domainRBAC, onboardedService.Namespace, onboardedService.Name, labels["app"])
-	//fmt.Println("convertCR looks like: ", convertedCR.ConfigMeta.CreationTimestamp)
-	expectedCR := getExpectedCR();
-	assert.Equal(t, expectedCR, convertedCR, "converted authz policy should be equal")
-}
-
-func getExpectedCR() model.Config{
-	var out model.Config
-	schema := collections.IstioSecurityV1Beta1Authorizationpolicies
-	createTimestamp, _ := time.Parse("", "12/8/2015 12:00:00")
-	out.ConfigMeta = model.ConfigMeta{
-		Type:      schema.Resource().Kind(),
-		Group:     schema.Resource().Group(),
-		Version:   schema.Resource().Version(),
-		Namespace: "test-namespace",
-		Name:      "onboarded-service",
-		CreationTimestamp: createTimestamp,
-	}
-	out.Spec = &v1beta1.AuthorizationPolicy{
-		Selector: &workloadv1beta1.WorkloadSelector{
-			MatchLabels: map[string]string{"svc": "productpage"},
-		},
-		Rules: []*v1beta1.Rule{
-			{
-				To: []*v1beta1.Rule_To{
-					{
-						Operation: &v1beta1.Operation{
-							Methods: []string{
-								"GET",
-							},
-						},
-					},
-				},
-			},
-			{
-				From: []*v1beta1.Rule_From{
-					{
-						Source: &v1beta1.Source{
-							Principals: []string{
-								"*",
-								"test.namespace/ra/test.namespace:role.productpage-reader",
-							},
-						},
-					},
-					{
-						Source: &v1beta1.Source{
-							RequestPrincipals: []string{
-								"*",
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-	return out
-}
-
-func getFakeDomain() zms.SignedDomain {
-	allow := zms.ALLOW
-	timestamp, err := rdl.TimestampParse("2019-07-22T20:29:10.305Z")
-	if err != nil {
-		panic(err)
-	}
-
-	return zms.SignedDomain{
-		Domain: &zms.DomainData{
-			Modified: timestamp,
-			Name:     DomainName,
-			Policies: &zms.SignedPolicies{
-				Contents: &zms.DomainPolicies{
-					Domain: DomainName,
-					Policies: []*zms.Policy{
-						{
-							Assertions: []*zms.Assertion{
-								{
-									Role:     DomainName + ":role.admin",
-									Resource: DomainName + ":*",
-									Action:   "*",
-									Effect:   &allow,
-								},
-								{
-									Role:     DomainName + ":role.productpage-reader",
-									Resource: DomainName + ":svc.productpage",
-									Action:   "get",
-									Effect:   &allow,
-								},
-							},
-							Modified: &timestamp,
-							Name:     DomainName + ":policy.admin",
-						},
-					},
-				},
-				KeyId:     "col-env-1.1",
-				Signature: "signature-policy",
-			},
-			Roles: []*zms.Role{
-				{
-					Members:  []zms.MemberName{username},
-					Modified: &timestamp,
-					Name:     DomainName + ":role.admin",
-					RoleMembers: []*zms.RoleMember{
-						{
-							MemberName: username,
-						},
-					},
-				},
-				{
-					Members:  []zms.MemberName{"productpage-reader"},
-					Modified: &timestamp,
-					Name:     DomainName + ":role.productpage-reader",
-					RoleMembers: []*zms.RoleMember{
-						{
-							MemberName: username1,
-						},
-					},
-				},
-			},
-			Services: []*zms.ServiceIdentity{},
-			Entities: []*zms.Entity{},
-		},
-		KeyId:     "colo-env-1.1",
-		Signature: "signature",
-	}
-}
